Fix comment about *code-coverage-info*.
*code-coverage-info* uses +code-coverage-unmarked+, not nil, to
indicate the state of a path that has not been visited. The
instrumentation generating code marks those records as constant to
ensure that multiple records for the same location get coalesced (the
instrumentation code subsequently modifies them anyways, suppressing
the warning about modifying a constant value). Because of that, it's
important that well-behaved user code won't contain any constant
lists that are coincidentally equivalent to (cons path state), where
path is a list of fixnums.

(cracauer commit on behalf of Sam)