compiler/ir1opt: lp#1563127: Account for DX when deleting vestigial exits
  * The original analysis of MVL+NLX issues causing stack analysis
to screw up in a variety of ways was largely accurate, but missed
the important case of the compiler wanting to nip an unknown-value
stack packet out from under a dynamic-extent stack packet, which
the %NIP-VALUES IR2-CONVERT optimizer rightly complains about.

  * The situation is caused by coalescing the LVARs for a
vestigial exit when there are USEs of the resulting ENTRY LVAR in
different dynamic-extent contours.

  * So, to fix, extend MAY-DELETE-VESTIGIAL-EXIT to verify that no
dynamic-extent leaves (variable or function) are visible in the
exit LEXENV without also being visible in the entry LEXENV.  Not
the most aggressive solution possible (as there should be no
problem so long as all USEs of the ENTRY LVAR have the same set of
visible dynamic-extent leaves), but a comfortably conservative
solution.