SAVE-LISP-AND-DIE calls REPACK-XREF (via TUNE-IMAGE-FOR-DUMP)
Since the most frequently referenced names (even just within SBCL
itself) are not known until the system is built, computing these
frequencies for sbcl.core has to be done around warm load time.

Since all xref information has already been collected and stored by that
time, all xref information has to be unpacked and repacked, this time
exploiting the compact encoding for the most frequent names.

Having to do this immediately before dumping sbcl.core, we might as well
always do it as part of SAVE-LISP-AND-DIE. That said, significant space
savings are unlikely to happen for subsequent dumps since the relative
frequencies of referenced names do not seem to change that much when
additional code is loaded (depends on the code, of course).

Some of the infrastructure is now shared with the sb-introspect contrib.