More compact xref storage
The previous storage format, a vector of vectors of the form

  #(#(NAME11 NUMBER11 NAME12 NUMBER12 ...) #(NAME21 NUMBER21 ...) ...)

used a vector header and five words for the vector plus two words for
each stored reference.

The new storage format uses a vector containing one octet vector and a
variable number of names:

  #(OCTET-VECTOR NAME1 NAME2 ...)

where the octet vector contains form numbers and indices into the stored
names as variable-length encoded integers. In many cases, the name index
and form number can be stored in a single octet.