x86-64: Allow CALL of a fixup.
This was disallowed allegedly to avoid "silently generating bogus code."
But there are reasons for wanting this. If you want to prohibit it because
there are ways in which it might be improperly used, the place to do that
is not in the only thing that knows how to encode the instruction.
Anyway the JMP emitter allows it, as does genesis.