Use bitmap_scan() when layout->bitmap is a bignum
Though bitmap_scan expects to deal with a range of memory whose size is a
multiple of N_WORD_BITS, it never accesses memory not indicated by a 1 bit
in the bitmap. Therefore this works agnostically of layout->length.

Aside from the improved algorithm, we only need to check for a forwarded
bignum once per scavenged object, not once per slot per object.