Add :compact-instance-header feature for x86-64.
This is part 2 of 4 of the overall immobile space feature.
The immobile allocator places all layouts below 4GB so that
a 32-bit pointer suffices to address them, which in turn allows
stuffing a layout pointer into an instance header.
In practical terms this yields either a 0-word or 2-word reduction
in the number of slots, depending on presence of trailing padding.

The feature is enabled by default on supported platforms.