mips: Prevent XEP-ALLOCATE-FRAME from trashing NARGS
  * Yet another entry in the saga of "XEP-ALLOCATE-FRAME vs. the
NARGS".  This has not been linked to any known failure on MIPS,
but was found by desk-checking all of the backends for this bug.

  * This, in theory, should be the last backend vulnerable to this
bug.  HPPA, SPARC, and ARM actually ran into it, PPC and MIPS were
found through desk-checking, ALPHA has NARGS as a descriptor-reg,
ARM64 doesn't need the temporary, and the x86oids don't maintain a
CODE register.