Return lispobj from forwarding_pointer_value(), fix trans_code().
forwarding_pointer_value() returns a tagged descriptor, so nearly all
uses would cast the result from lispobj* to lispobj. It's more clear
to declare it with the right type in the first place.
In particular, you can not dereference in C without untagging.

Having observed that, it was apparent that trans_code() sometimes
returned tagged pointers, but usually not. Through a small miracle,
this was harmless - callers typically ORed in a lowtag anyway.
Nonetheless, we should always return a naturally aligned pointer.