Increase the amount of debug-fun info generated by default.
Generate full debug info when COMPUTE-DEBUG-FUN is 1 or greater, this
increases core size on x86-64 by 7MB but is well worth it.
COMPUTE-DEBUG-FUN=0 now means "minimal", the old "no" wasn't saving
that much space, and it's easier to get the old behavior by doing
(declare (optimize (sb-c::compute-debug-fun 0))).

Change the default value of *DISASSEMBLE-ANNOTATE* to NIL, with more
debug info it tries to annotate stuff more often and fails more often
and is just as useless as before.