Fix GENTEMP for simultaneous use by multiple threads.
Instead of unsynchronized calls to FIND-SYMBOL and INTERN,
just use INTERN's second value to see if it created something.
It's still a bad idea to use this deprecated function,
but at least it does what it promises.

Also, FLET-ize the helper code, and remove redundant decls
conveyed through cl-specials.lisp and/or fndb.lisp.