Add test case for WITH-PINNED-OBJECTS
  * In commit 9d6b6ac28d8db6c536842d2f07f7e8c1e2ff5a31, Douglas
Katzman mentioned something which "... might confirm what CSR
hinted at which is that neither evaulator does W-P-O correctly."

  * CSR was right: Neither evaulator does W-P-O correctly... on
x86oids.  The mechanism is less likely to fail on other backends,
for various reasons, but the least we can do is actually test it,
so that we know if there's a problem or not.

  * Tested on x86-64/Linux, with sb-eval and sb-fasteval, in both
evaluate and compile evaluator modes.