ir1opt: Better D-X contour detection for vestigial exit deletion
  * Apparently, special variables aren't leaf-dynamic-extent, even
when they're dynamic-extent bound, so the logic for detecting the
D-X contours broke.

  * Crib from PHYSENVANAL's approach for deciding if cleanups need
inserting, since it's the inserted cleanups that cause some of the
proximate damage (even if it's the exit-deletion that causes them
to be wrong).

  * Clearly, I haven't been fixing these stack analysis / vestigial
exit bugs, merely driving them to become more subtle.  With a bit
of luck, maybe this time it's really fixed.

  * Fixes lp#1655011.