ir1opt: Better D-X contour detection for vestigial exit deletion... again
  * So, while a given deleted EXIT can only have one successor, if
it is the only node in its block it can have multiple predecessors,
meaning that the vestigial exit CAST can end up with multiple
predecessors.  Additionally, if one of the predecessors is itself
an EXIT from another CLAMBDA, it would have no cleanups, so we
must check all of our predecessors for DX cleanups, not just the
first (and presumed only) one.

  * And, in the spirit of "driving the bug to become more subtle",
I fully expect there to be at least one more way to break this
logic.