ir1opt: Better vestigial exit deletion logic... yet again
  * One new failure test case, (:EXIT-DELETION :2017-01-30
:EVEN-MORE-SUBTLE), which broke due to none of the predecessor
blocks of the vestigial exit cast having the appropriate set of
cleanups in force.

  * One new test case to cover the need for checking the TAGBODY
ENTRYs in vestigial exit deletion.  They don't tend to crash the
compiler when they go wrong, "merely" generate bogus code.  Which,
really, is worse: The compiler crash is hard to miss, the bogus
generated code a lot easier.

  * And a fix and simplification for MAY-DELETE-VESTIGIAL-EXIT.
"This time for sure!"