x86-64: New call convention for #!+immobile-code
For those keeping score, this is part 4 of 4 of the immobile space feature,
but there's more to follow. Example:

* (disassemble '(lambda () (eval (read))))
; disassembly for (LAMBDA ())
...
; 719:       E81AF4C2FD       CALL #x2032CB38                 ; #<FDEFN READ>
; 71E:       B902000000       MOV ECX, 2
; 723:       FF7508           PUSH QWORD PTR [RBP+8]
; 726:       E9CD7FC3FD       JMP #x203356F8                  ; #<FDEFN EVAL>

RAX is not loaded prior to CALL, and we don't use the constant from the
code header for anything except to preserve liveness. This reduces pressure
on the data TLB and optimizes for the common case of a named call to
non-generic functions. GFs and closures need a little help, as they get
wrapped inside a tiny piece of code that loads RAX prior to bouncing.

Performance measurement suggests that the instruction decoder and iTLB
are able to deal with this better than multiple data indirections.
It is a tiny step to statically link all calls by removing the indirection
through the fdefn for named call if it is neither to a closure nor GF.