Consolidate the nine? three? two? implementations of WITH-PINNED-OBJECT
  * At first blush, this looks to be three primary variations
spread over nine backends, but in actual practice one of the
variations (the cheneygc-only version) is a strict subset of
another, so there's only two "real" versions to contend with.

  * And since it's essentially purely duplicated code (a backend
had either the alpha-hppa-mips version, the arm-arm64-ppc-sparc
version which is the same as the former when built for cheneygc,
or the x86-x86-64 version), it should be generic.  Besides, it's
one less thing to change when we implement threads for MIPS, HPPA,
or Alpha.  Stop laughing.

  * It may-or-may-not make sense to convert one or both of the
x86oid backends to using *PINNED-OBJECTS*, but that's a question
for another time.