Add support for saving executables in ELF format
Adds a new feature :SB-ELF-CORE that depends on SB-LINKABLE-RUNTIME.

All GC spaces are saved in a pair of ELF sections: one that maps the
code from the executable and one that represents the empty hole in that
space and is zero-filled. Currently, these are:
 * .sbcl_readonly & .sbcl_readonly.zero
 * .sbcl_static & .sbcl_static.zero
 * .sbcl_immobile.fixedobj & .sbcl_immobile.fixedobj.zero
 * .sbcl_immobile.varyobj & .sbcl_immobile.varyobj.zero
 * .sbcl_dynamic & .sbcl_dynamic.zero

Runtime options are saved in a struct that is bound by default to a weak
symbol named "linked_in_core_data" and placed in the section ".rodata".
That way, the non-ELF sbcl runtime gets the default version and the ELF
executables get the one from the ELF core. This also acts as an easy way
for the runtime to determine that its GC areas have are already
allocated by the runtime linker.

A new flag :C-LINKABLE-CORE is added to SAVE-LISP-AND-DIE which has it
call save_elf_core() in the C runtime. That function creates two files:
the core as an ELF object file and a linker script (with ".ldcmd"
appended to the name of the core file), that pins every ELF section to a
fixed address at startup.

Usage:

$ sbcl --eval '(sb-ext:save-lisp-and-die "core.o" :c-linkable-core t)'
$ cc -g -Wl,--export-dynamic -Wl,@core.o.ldcmd \
    -L$SBCLDIR/src/runtime/libelf $SBCLDIR/src/runtime/sbcl.o core.o \
    -ldl -lpthread -lm -lelf_private \
    -o sbcl_with_elf

Based on a patch by James Knight