Use ordinary funcall for GFs made by ENSURE-ACCESSOR
At least on x86-64 with static linking of GFs,
(FUNCALL #'(SOME-NAME) ...) is better than
(FUNCALL (LOAD-TIME-VALUE (FDEFINITION '(SOME-NAME))) ...)
Other backends are indifferent to this change.

Additionally, the side-effectful part of the L-T-V form needn't be
dumped into fasls. Instead the concept of an fdefn which always has
a functional binding can be stuffed into find-or-create-fdefn.