Remove a kludge, fix a problem in UNSEEN-THROW-TAG handler.
Switching two files around allows inlining SB!DI:FRAME- accessors.
This revealed the bug that "(WHEN FRAME" should have been "(WHEN DOWN"
as the compiler was kind enough to point out that "NIL is not a FRAME".

Additionally, we can, and must, remove a line from GC-REINIT due to
a forward referenced special - now a global - which is nicer anyway.