Rename %SIMPLE-FUN-SELF vops, make closures great again.
The only reason for calling %SIMPLE-FUN-SELF is to extract the 'fun' value
from a closure as a function object, which happens for free on non-x86.
So "naming is hard", but seriously? It's just %CLOSURE-FUN.

Also a drive-by fix for build breakage: other than PPC, non-x86 platforms
failed due to style-warnings. A source-transform can generate a warning about
the function that it replaces, if that - %CLOSURE-CALLEE in this case - was
never defined. The PPC build allows warnings for "reasons". Ugh.