Cut down the size of 'struct page' for 64-bit words.
* Move the 'need_to_zero' flag into the low bit of the
  'bytes_used' field, and mask it off when reading.
* Pack the 'has_pin_map' field into the 'flags'.
* Use 4 bytes to hold 'scan_start_offset', scale it,
  and hide the scaling in a preprocessor macro.

The tricky piece here is that if the scaled scan_start_offset
is still too small to store its full value, we have to look
back for an earlier page which is the start of a region.
This is unlikely though - a single 32 TB region is probably rare.

All other things being equal, space is time, and despite extra
code to compute the hairy case of scan_start_offset, this yielded
around .5% speedup. It might work even better the larger the heap.