Fix test failure as a result of #+immobile-code.
Normally the trampoline made by INSTALL-GUARD-FUN is a closure that captures
SYMBOL, invoking ERROR to construct an UNDEFINED-FUNCTION condition.
Immobile-code works differently though: to store a closure as a named function
requires consing a new trampoline to call the closure, since we don't load RAX
with the address of the callee. But we don't cons anything to install the
undefined-tramp asm routine. We use the CALL instruction to get from the fdefn
to the asm routine, which pops the return PC to compute the fdefn, and it
invokes undefined-fun-error with an fdefn which has a possibly wacky name,
if it named a special form, in which case we have to unwacky-ize it.