Optimize test for unbound CLOS slots, part 2 of 2.
Change unbound slot marker to the unbound marker object, among the benefits
being that a slot intended to contain a symbol could accidentally have stored
SB-PCL::..UNBOUND-MARKER.. but should not store the unbound marker; and
comparison is 4 bytes shorter on x86-64, 3 bytes shorter on x86,
and on all others avoids a code header constant.

This change is nearly invisible, even to code using SB-PCL:+SLOT-UNBOUND+
as recommended in doc/manual/behond-ansi.texinfo. There are a few observable
differences which should not matter to production code:
1. (CAS (SLOT-VALUE obj 'slot) sb-pcl:+slot-unbound+ 'newval)
   might return the unbound marker instead of a "genuine" object
   in the language. This is fine - the only correct use of the CAS result
   is to compare again the expected old value.
2. evaluating SB-PCL:+SLOT-UNBOUND+ at the REPL prints
   "EVAL returned an unbound marker."
   The test was inherited from CMUCL, and it's not clear why
   that particular test for the unbound marker is sensible.
   Arguably a valid fix would be silently allow '* become unbound,
   and thus print "#<unbound marker {51}>" which would happen in any other
   situation wherein an unbound symbol's value was inadvertently obtained.