x86-64: Improve symbol-value even without #!+immobile-symbols.
Better code is generated for #!+immobile-space. The problem is determining
which symbols are immobile. Since the cross-compiler always uses immobile space
for symbols when it can, at least that much is easy to decide. The other case
is compilation to memory: if a symbol has a low address, it won't move.

Unfortunately, compiling to fasl can not make use of this, because after
deciding to use a 4-byte immediate operand to reference a symbol,
we'll lose badly if the loading process has the symbol at a high address.