Fix compilation of &OPTIONAL binding with non-EQL-comparable constant default.
Uses of "hairy" named constants (as from SB-INT:DEFCONSTANT-EQX) should be
compiled into LOAD-TIME-VALUE on the respective symbol. Most were, except that
&OPTIONAL arg defaults would eagerly evaluate the symbol at compile-time.
In addition to causing a semantic problem for vectors, the compiler complained
about absence of a MAKE-LOAD-FORM method on a structure class where no such
literal object ought to be dumped to the fasl at all.