Add some bits to LAYOUT
Without looking at layout-classoid we can test whether an object is
an instance of STRUCTURE-OBJECT, CONDITION, or STANDARD-OBJECT.

In addition to being slightly faster to test for those types in EQUALP
and COERCE-TO-CONDITION, it's needed for compact instance headers
on 32-bit word size. The issue there is that CONDITION instances are
variable in length, where length is not conveyed by the layout.
The other two metatypes have the property that layout implies length.