Stop "initializing" in triplicate the GC control symbols
create_thread_struct() touches a bunch of Lisp variables.
The DEFVAR initforms for those are irrelevant, so remove them.
Assignments in cold load are not needed, so don't do them.
The #ifdefs to decide what to initialize are all gone.

Moreover, if multi-threaded, then all these symbols
are globally unbound, but thread-locally bound.
Once-and-only-once, for the win.