Fix inclusions, add remark about what else needs repair
gencgc-internal was included twice from gc-internal,
first from 9f10bc102a by Dan Barlow and then 40f6a8f39d by Nikodemus.

Neither place was the best, as it was not possible for inline functions
in gencgc-internal or cheneygc-internal to use gc_abort()
which was #defined only after inclusion of the GC-specific internals.

And linux-os should not need an #ifdef for which GC to include