Resolve logical disparity in binding stack scavenging
TLS entries do not scavenge a symbol, so the binding stack is not strictly
speaking a GC root. This change makes non-thread builds also not treat the
binding stack as a GC root for its symbols. But any transported symbols
will be fixed up in a post-scavenging step in the manner of a weak relation.

This ensures that if there were a scenario in which a symbol would have
been garbage were it not for presence on the binding stack, presence or
lack of threads will have similar behaviors (i.e. potentially failure).