Reduce #ifdefage and make verify_space() more rigorous on all backends
Generally 'fdefn->raw_addr' should be converted to a function descriptor
before using it for anything other than function call.
verify_space() failed to convert the stored fixnum to a pointer.
(Only sparc and ARM store that slot directly as a simple-fun)

We also fail in MAP-REFERENCING-OBJECTS, but that's broken for other
reasons - the *IGNORE-AFTER* variable randomly skips referencing objects.