Reserve 1 bit in almost all object headers for a GC mark bit.
- Vectors don't use the header for their length, so we can easily take 1 bit.
- Most other objects don't need the full number of bits in the header.
- Bignums are the exception; a 32-bit header uses 24 bits for bignum-length
  which when added to 8 for the widetag leaves no extra bit.

This change does not ensure that Lisp code all works if an object's mark bit
is on; it just ensures that if a sizetab[] function is invoked on a marked
header, the correct size is obtained.

It will take a ton of work to implement a concurrent marking strategy,
but pending a few easy changes, we'll be able to sniff out tenured garbage.