gencgc: Zero-fill tenured (generation 6) objects under user control.
"Object permanence in gen6 is a feature" said * NOBODY EVER *

The reason gen6 is un-freeable is that there is no means of evicting dead
objects from pages containing live objects, because to keep an object live
is synonymous with copying it; but we don't want to pay a price for copying
all objects in the image on every full GC cycle.

As a partial solution, invoking (GC :GEN 7) will zero out pointers in
unreachable pseudo-static objects. Only generation 6 objects that are reachable
from truly static objects (or thread stacks) are considered reachable.
This makes no memory available for re-use, causes no promotions to occur,
and compacts nothing. So what good is it?

Answer: If you have a known problem with inadvertent object retention due to
the collector's insistence that generation 6 contains only immortal roots,
you can cure that problem by wiping out the pointers as a mitigation.
This is a common problem with vectors that were saved in the image,
attached to a defglobal that subsequently got assigned a new vector.
The hack that is customarily employed within SBCL to manually zero-fill
old vectors does not lend itself to vectors that participate in wait-free
algorithms. (It is impossible to write user code that detects absence of
concurrent readers; if you could do that, it wouldn't be wait-free)

Besides fixing accidental retention, there are other interesting possibilities:

- writing a "tree shaker" that eliminates transitive reachability of garbage
  without needing to kill the process to disover whether you shaker
  actually did what it was supposed to. i.e. even though there is a way
  to do this, it's nicer that process death isn't the only way to do it

- could allow decoupling save() from gc() in gc_and_save()

- could potentially form the basis of a real concurrent marking strategy