Increase granularity of core page alignment
The entry->address divisor is mostly arbitrary, but has to be at least
as fine as the alignment resolution of immobile pages to support relocation.
Suppose that the OS maps immobile space at 0x50001000 on a particular run,
and the user saves that core. The address must be exactly that,
not constrained by the possibly larger value of +BACKEND-PAGE-BYTES+.