Pin page-spanning small objects on an individual basis
Page-spanning non-large-page objects used to pin all the small objects that
they shared a page with, because wipe_nonpinned_words was unable to deal
with them. So instead we would mark the pinned object's pages as dont_move,
and rely on the coarser-grained pin.  We don't have to do that any more.
Also, preserve_pointer is vastly simpler.

But this revealed a bug with my previous optimization
of page_ends_contiguous_block_p(). We do need to test whether a page is
incompletely full, but we don't need all 5 tests that it performed.