regalloc: Faster hiding/showing of vertex neighbors
In COLOR-NO-CONFLICTS-P, iterating over an oset and checking for
invisibility was the main bottleneck. Maintain a SIMPLE-VECTOR of
currently incident neighbors and iterate over that instead.

This makes the invisibility logic unnecessary.

This is part of a series of changes that reduce the amount of work
done in COLOR-NO-CONFLICTS-P which is the hottest part of the
iterative register allocation method.