Fix source of inefficiency in trans_code()
trans_code invoked gc_general_copy_object instead of copy_large_object.
The general routine blindly copies anything given to it, even though both
the old object and its replica are on pages unshared with other objects.

Additionally the comment in general_copy_large_object that
"If it's large enough then it's copied to a large object region"
was wrong, because it never upgrades small objects to large.