Mostly revert 2df2efd246
That patch caused there to be more page-spanning objects, which is ok for
code components (they are rarely touched) but not so great for other objects,
because our algorithm for small-object pages forces more scavenging than
desired- it can only stop on a page whose successor has scan_start_offset 0.
Some improvements to that are pending.