Fix stack exhaustion in with-alien on non-x86oids.
The code claimed that using UNWIND-PROTECT to restore NSP wasn't
required because it is restored during non-local exits, true, but
local exits do not restore the dynamic environment.
UNWIND-PROTECT is a bit expensive and since NSP is already restored on
NLX, it can behave just like special bindings. Introduce a new special
form, RESTORING-NSP, which does just that.