Delay choosing representation for constants.
Previously, some heuristics were employed to select between boxed and
unboxed represenations for constants at IR2 conversion time. But they
often selected the wrong representation, because IR2 conversion time
is too early to decide. Normal TNs are assigned their SC in
SELECT-REPRESENTATIONS, with a little adjustment it works on constants
as well.
MAKE-CONSTANT-TN leaves the SC unassigned if a constant can either be
boxed or unboxed (which is floats and complex floats).