Generalize immobile space addresses
This is the first of several changes to split immobile space into two possibly
discontiguous ranges. Certain assumptions are removed,
but only in as much as they pertain to naming convention:
(1) that *immobile-space-free-pointer* is the upper bound on both subspaces
(2) that varyobj-space-start is fixedobj-space-start + fixedobj-space-size

Those constraints are not actually removed as yet. A consequence of this patch
is that GENCGC-SPACE-SETUP changed a lot, but hopefully the keyword arguments
make more sense now.
Corrected a fencepost error in IMMOBILE-SPACE-ADDR-P.