make upgraded-complex-part-type less clever
Unfortunately the requirements under CLHS TYPEP constrain the
implementation of UPGRADED-COMPLEX-PART-TYPE severely; the
requirement that all others of the same specialized representation
satisfy the typep requirements individually and severally on real
and imaginary parts means (we believe) that the only possible
implementation of UPGRADED-COMPLEX-PART-TYPE is effectively the
trivial implementation.  Make it so.