Don't note about deleting unreachable casts.
CASTs are just value filters and do not represent code and they can be
moved around making CASTs from the original source code appear in code
inserted by the compiler, generating false deletion notes.
And if a block with the original source gets deleted the node that
produces the value for the CAST will get a note, no need to note
twice.