x86-64: Prefer RIP-relative mode for accessing assembly tramps
More work to defrag, but less work to relocate spaces
(in reality there are < 1 dozen affected instructions)

However, the win is that intra-varyobj-space opaque pointers (ones not seen
by GC) are all space-relative, so as long as fixedobj space does not need
relocation, the varyobj space can move anywhere with no fixups internal to
that space. We must fix pointers into it of course.