safepoint: Fix deferrable interrupt handling while phase is not GC_NONE
  * Because we no longer use pseudo-atomic-interrupted to trigger
GC, check_interrupt_context_or_lose() needs to have some way to
know that a GC is, in fact, going to happen, lest it see that a
GC (or stop the world) is pending without some reason why.

  * There's a small amount of cleverness here to make sure that we
don't take the gc state lock when we might already have it (and
it's not like we actually care what the phase is in these cases).