safepoint: Partial restoration of PPC support
  * "Partial" for two reasons: First, this only works for systems
with a 4k page size, due to the CSP trap pages being presumed to
be 4k.  And second, it's not stable, as expected for something
with "I saw it almost pass tests once.  Ship it." in the inaugural
commit message (fd8e8143cf02ac767e2a46a2bc526933e68ef583).  That
said, it now survives the build process and can run the test suite
without any especially egregious failures.

  * First off, we "fix" lisp_memory_fault_error() to not try to
block GC signals on safepoint.  This will still likely break, but
at least it's not causing an outright build failure anymore.

  * Next, fun and games with the GSP trap page.  The Linux ABI
allows both 4k and 64k page sizes (and possibly others?), so our
backend page size is 64k.  This puts GC_SAFEPOINT_PAGE_ADDR at
just over 64k before NIL.  The safepoint instruction is an LWZ
relative to reg_NULL, but uses a 4k (plus the distance from the
start of static space to NIL's tagged representation) offset,
probably having been written before the 64k page size thing was
known.  We can't hit the base of the safepoint page, so introduce
a new lisp-side constant for the (negative) offset from the start
of static space to the actual trap address, export a derived
symbol to C for the trap address (no sense computing it C-side),
use the derived constant in our trap emission function, fix up
handle_safepoint_violation() to check for the trap address, and
we're (not) done.

  * More fun and games with the GSP trap page.  Only requesting to
map four bytes doesn't help when your host page size is smaller or
equal to the distance from the page address to the trap address.
Use the actual defined backend page size instead.

  * And while we're altering the GSP trap mapping, it doesn't need
to be writable.  We don't write to it.  We don't WANT to write to
it.  And now we don't map it as writable.