Remove gencgc_partial_pickup
This flag was basically useless.  The logic in gencgc_pickup_dynamic
was made overly complicated to support a usage that happens exactly
one time after genesis, that of not having a page-table in the core,
just so that gc_search_space() might run a little faster.
After genesis time, it can't happen that there is no page table.
So remove that needless complexity. Platforms which have precise gencgc
almost never call gc_search_space(), so there is especially no worry.

Also, change the core header entry to directly represent the number
of page-table pages stored rather than writing that value onto the
first page of data holding the PTEs.