Remove constant scale factor from address in struct ndir_entry
The system linker would have to know how to divide by an arbitrary
value if we want the linker to write that field using a relocation.

It could be argued that 32-on-64 platforms need a multiplier if lisp words
are 32-bits while addresses are 64. However, the only such backend, Alpha,
does not allow dynamic space above 4GB.

Also update editcore due to prior change