safepoint feature compatibility noise
  * :SB-SAFEPOINT is only implemented for ppc, x86, and x86-64.
And :SB-SAFEPOINT-STRICTLY requires :SB-SAFEPOINT.  Add these to
the feature compatibility tests so that the build will bomb out
early (with an explanation of why) instead of taking forever to
discover a build-configuration mistake.

  * And both features affect generated code, and therefore the
FASL file format.  Add them to the corresponding list so that
anyone who builds the same SBCL version with various options (such
as an SBCL maintainer might) don't get bitten unnecessarily by
running safepoint code on a non-safepoint system or vice versa.