Use BOXED-CONTEXT-REGISTER in several places in debug-int
  * The theory being that we already know that it's boxed, so we
don't need to pay the price of a heap search for MAKE-LISP-OBJ,
and the WITHOUT-GCING gets pushed to inside BOXED-CONTEXT-REGISTER
for the time being (and eventually it should get removed entirely,
but I'm taking this one step at a time for the moment).

  * Non-x86oid FIND-ESCAPED-FRAME / CODE-OBJECT-FROM-BITS looks
like a candidate for this as well, but it turns out that there are
more fundamental issues there, worthy of a commit (or series of
commits!) of their own.